# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Recogno API",
    "description": "Backend for Recogno, a spaced-repetition trainer for DSA and system-design pattern recognition.\n\nThe drill loop is: `GET /drill/next` for a problem stripped of its answer, `POST /drill/submit` to grade a guess and advance its FSRS card, `GET /drill/due-count` for the reps-due badge.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from POST /auth/login, /auth/register or the Google callback. Paste just the token — Swagger adds the \"Bearer \" prefix."
      }
    },
...
  },
  "paths": {
...
    "/auth/google": {
      "get": {
        "summary": "Begin Google sign-in",
        "tags": [
          "auth"
        ],
        "description": "Redirects to Google. Returns 501 when the server has no Google credentials — call `GET /auth/providers` first to find out.",
        "security": [],
        "responses": {
          "302": {
            "description": "Default Response"
          },
          "501": {
            "description": "Something the caller asked for does not exist or is malformed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Something the caller asked for does not exist or is malformed."
                }
              }
            }
          }
        }
      }
    },
    "/auth/google/callback": {
      "get": {
        "summary": "Google sign-in callback",
        "tags": [
          "auth"
        ],
        "description": "Exchanges the authorization code for a Recogno session. Redirects to OAUTH_SUCCESS_REDIRECT when set, otherwise returns the session as JSON.",
        "security": [],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "displayName": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "avatarUrl": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "emailVerified": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "displayName",
                        "avatarUrl",
                        "emailVerified",
                        "createdAt"
                      ],
                      "additionalProperties": false
                    },
                    "accessToken": {
                      "type": "string",
                      "description": "Send as `Authorization: Bearer <token>`."
                    },
                    "refreshToken": {
                      "type": "string",
                      "description": "Opaque. Exchange at POST /auth/refresh; it rotates on every use."
                    },
                    "tokenType": {
                      "type": "string",
                      "enum": [
                        "Bearer"
                      ]
                    },
                    "expiresIn": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991,
                      "description": "Access token lifetime in seconds."
                    }
                  },
                  "required": [
                    "user",
                    "accessToken",
                    "refreshToken",
                    "tokenType",
                    "expiresIn"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "302": {
            "description": "Default Response"
          },
          "400": {
            "description": "Something the caller asked for does not exist or is malformed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Something the caller asked for does not exist or is malformed."
                }
              }
            }
          },
          "501": {
            "description": "Something the caller asked for does not exist or is malformed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Something the caller asked for does not exist or is malformed."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}